'betty'

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -11,7 +11,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	 long unsigned int i;
+	unsigned int i;
 
 	if (array == NULL || action == NULL)
 		return;
